Added fixes after first code review

+++ Blog.py
@@ -1,5 +1,6 @@
 from model.Post import Post
 from model.User import User
+from model.Like import Like
 from model.Comment import Comment
 import webapp2
 import os
@@ -15,18 +16,19 @@ class BlogFront(BlogHandler):
     """ Render blog's front page """
     def get(self):
         """ Front page's GET render """
-        posts = db.GqlQuery("Select * from Post order"
-                            "by created desc limit 10")
+        posts = db.GqlQuery("""Select * from Post order
+                            by created desc limit 10""")
         username = self.user and self.user.name
+
         self.render("front.html", username=username, posts=posts)
 
 
 class PostPage(BlogHandler):
+    """ Render blog's post """
     def get(self, post_id):
+        """ Render blog's post - get method """
         key = db.Key.from_path('Post', int(post_id), parent=utils.blog_key())
         post = db.get(key)
-
-        comments = post.comments
 
         if not post:
             self.error(404)
@@ -37,12 +39,13 @@ class PostPage(BlogHandler):
         self.render("permalink.html",
                     username=username,
                     post_id=post_id,
-                    comments=comments,
                     post=post)
 
 
 class DeletePost(BlogHandler):
+    """ Delete post """
     def get(self, post_id):
+        """ Delete post - get method """
         key = db.Key.from_path('Post', int(post_id), parent=utils.blog_key())
         post = db.get(key)
 
@@ -50,27 +53,33 @@ class DeletePost(BlogHandler):
             self.error(404)
             return
 
-        post.delete()
-
-        self.redirect("/blog")
+        if post.author:
+            post.delete()
+            self.render('delete.html',
+                        message="Post succesfully deleted")
+        else:
+            self.render('delete.html',
+                        message="Error attempting to delete post")
 
 
 class LikePost(BlogHandler):
+    """ Like post class """
     def get(self, post_id):
+        """ Like post - get method """
         if self.user:
             key = db.Key.from_path('Post',
                                    int(post_id),
                                    parent=utils.blog_key())
             post = db.get(key)
+            author = User.by_name(self.user.name)
+            like = True
 
             if not post:
                 self.error(404)
                 return
 
-            post.likes = 0 if post.likes is None else int(post.likes)
-            post.likes += 1
-
-            post.save()
+            post_like = Like(author=author, post=post, like=like)
+            post_like.put()
 
             self.redirect('/blog/%s' % str(post.key().id()))
         else:
@@ -78,7 +87,9 @@ class LikePost(BlogHandler):
 
 
 class EditPost(BlogHandler):
+    """ Edit post class """
     def get(self):
+        """ Like post class - get method """
         post_id = self.request.get("key")
         key = db.Key.from_path('Post',
                                int(post_id),
@@ -95,6 +106,7 @@ class EditPost(BlogHandler):
             self.redirect('/blog')
 
     def post(self):
+        """ Like post class - post method """
         post_id = self.request.get("key")
         key = db.Key.from_path('Post',
                                int(post_id),
@@ -127,7 +139,9 @@ class EditPost(BlogHandler):
 
 
 class NewPost(BlogHandler):
+    """ New post class """
     def get(self):
+        """ New post class - get method """
         if self.user:
             self.render("newpost.html",
                         title="New Post",
@@ -136,31 +150,38 @@ class NewPost(BlogHandler):
             self.redirect('/login')
 
     def post(self):
-        subject = self.request.get("subject")
-        content = self.request.get("content")
-        author = self.user.name
-        error = ""
+        """ New post class - post method """
+        if self.user:
+            subject = self.request.get("subject")
+            content = self.request.get("content")
+            author = User.by_name(self.user.name)
 
-        if subject and content:
-            p = Post(parent=utils.blog_key(),
-                     subject=subject,
-                     content=content,
-                     author=author)
-            p.put()
-            self.redirect('/blog/%s' % str(p.key().id()))
+            error = ""
+
+            if subject and content:
+                to_post = Post(parent=utils.blog_key(),
+                               subject=subject,
+                               content=content,
+                               author=author)
+                to_post.put()
+                self.redirect('/blog/%s' % str(to_post.key().id()))
+            else:
+                error = "subject and content, please!"
+                self.render("newpost.html",
+                            title="New Post",
+                            username=self.user.name,
+                            subject=subject,
+                            content=content,
+                            error=error)
         else:
-            error = "subject and content, please!"
-            self.render("newpost.html",
-                        title="New Post",
-                        username=self.user.name,
-                        subject=subject,
-                        content=content,
-                        error=error)
+            self.redirect('/login')
 
 
 # Comment section
 class NewComment(BlogHandler):
+    """ New comment class """
     def get(self):
+        """ New comment class - get method """
         if self.user:
             post_id = self.request.get("key")
             key = db.Key.from_path('Post',
@@ -180,6 +201,7 @@ class NewComment(BlogHandler):
             self.redirect('/login')
 
     def post(self):
+        """ New comment class """
         if self.user:
             post_id = self.request.get("key")
             key = db.Key.from_path('Post',
@@ -191,17 +213,13 @@ class NewComment(BlogHandler):
                 self.error(404)
                 return
 
-            if post.key().id():
-                content = self.request.get("content")
-                author = self.user.name
-                error = ""
+            content = self.request.get("content")
+            author = User.by_name(self.user.name)
 
-                if content:
-                    c = Comment(post=post, content=content, author=author)
-                    c.put()
+            if content:
+                comm = Comment(post=post, content=content, author=author)
+                comm.put()
 
-                self.redirect('/blog/%s' % str(post.key().id()))
-            else:
-                self.redirect('/blog')
+            self.redirect('/blog/%s' % str(post.key().id()))
         else:
             self.redirect('/login')
